requirements.txt
tempdir
nose
mock
biopython==1.63
numpy==1.8.0
unittest2
logconfig
filehandle
python-coveralls 
matplotlib
future
